Move cli to `bin/`
