+++ bin/cli.js
@@ -7,8 +7,8 @@ import log from 'loglevel'
 import logPrefix from 'loglevel-plugin-prefix'
 import { Command } from 'commander'
 
-import { WACZ } from './index.js'
-import { PACKAGE_INFO, LOGGING_COLORS } from './constants.js'
+import { WACZ } from '../index.js'
+import { PACKAGE_INFO, LOGGING_COLORS } from '../constants.js'
 
 const program = new Command()
